\name{freesortphi-package}
\alias{freesortphi-package}
\alias{freesortphi}
\docType{package}
\title{
\packageTitle{freesortphi}
}
\description{
\packageDescription{freesortphi}
}
\details{

  For a complete list of functions, use \code{library(help =
    "freesortphi")}.

  All functions are concisely documented, use the help function e.g
  \code{?phi.intra}.

  For a brief tutorial on these methods, see the Appendix of Lawson et
  al. (2016).

}
\author{
\packageAuthor{freesortphi}

Maintainer: \packageMaintainer{freesortphi}
}
\references{
  Lawson, R., Chang, F. & Wills, A.J. (2016). Free classification of
  large sets of everyday objects is more thematic than
  taxonomic. \emph{Manuscript submitted to Acta Psychologica}.
}

\keyword{ package }

